Extract cell editors into their own classes
This will allow them to be used in other tables.
